Create ExcelUtility_CreateHeaders_And_InputDataEveryRun
This excel utility will create excel file, add the headers and enter the data in excel.
It will keep on adding data in rows after every run and will not delete the previous row. 
If the header already exist because of the previous run then it will not add the headers. 
The excel will be created under src/test/java/resources

You need to create a utils package under src/test/java and create ExcelWriter class there